+++ pyloggor/__init__.py
@@ -1,1 +1,4 @@
 from .pyloggor import pyloggor
+
+__version__ = "2.0.0"
+__all__ = ["pyloggor"]